refactor: colocate root quest dashboard tests
