+++ components/quests/quest-dashboard/index.realtime-delete.test.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { render, waitFor, screen, act } from "@testing-library/react";
-import QuestDashboard from "../quests/quest-dashboard";
+import QuestDashboard from "./index";
 import { useAuth } from "@/lib/auth-context";
 import { useRealtime } from "@/lib/realtime-context";
 import { supabase } from "@/lib/supabase";
